edit all elements and sort by controllers

+++ database/migrations/2020_07_29_063841_create_orders_table.php
@@ -27,7 +27,7 @@ class CreateOrdersTable extends Migration
             //podatki o statusu narocila
             $table->string('completed')->default('no');
             $table->string('shipped')->default('no');
-            $table->double('price')->default(0);
+            $table->integer('price')->default(0);
 
             //podatki o torti
             $table->string('napis')->nullable();
